backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.71.1
google-cloud-pubsub==2.23.1
google-cloud-bigquery==3.25.0
google-generativeai==0.8.3
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
geopy==2.4.1
numpy==1.26.4
